实现analysis，修改对应vo&po（to long）

+++ src/main/java/vo/admin/UserAnalysisDataVO.java
@@ -7,7 +7,7 @@ import java.util.Map;
  */
 public class UserAnalysisDataVO {
 
-	private int userNum;								//总的用户数量
+	private long userNum;								//总的用户数量
 	private Map<String, Integer> registerNumByTime;		//按时间获取注册数量
 
 	public UserAnalysisDataVO() {}
@@ -16,16 +16,16 @@ public class UserAnalysisDataVO {
 	 * @param userNum 总的用户数量
 	 * @param registerNumByTime 按时间获取注册数量
 	 */
-	public UserAnalysisDataVO(int userNum, Map<String, Integer> registerNumByTime) {
+	public UserAnalysisDataVO(long userNum, Map<String, Integer> registerNumByTime) {
 		this.userNum = userNum;
 		this.registerNumByTime = registerNumByTime;
 	}
 
-	public int getUserNum() {
+	public long getUserNum() {
 		return userNum;
 	}
 
-	public void setUserNum(int userNum) {
+	public void setUserNum(long userNum) {
 		this.userNum = userNum;
 	}
